consolidate contact methods
