+++ messages/fr.json
@@ -159,7 +159,7 @@
     "faq": {
       "title": "Vous vous demandez peut-être...",
       "subtitle": "Ça n’existe pas déjà ? Quelle est votre mission ? Vous avez des questions, nous avons, espérons-le, des réponses.",
-      "footer": "Encore des questions ? <page>Nous sommes là pour vous aider</page>."
+      "footer": "Encore des questions ? <page>Contactez-nous</page>."
     }
   },
   "Partners": {
@@ -571,9 +571,9 @@
     "body": "Désolé, nous n’avons pas trouvé la page que vous cherchiez."
   },
   "Support": {
-    "title": "Aide",
-    "metaDescription": "Réponses et options de contact pour obtenir de l’aide avec Peels.",
-    "subtitle": "Trouvez des réponses aux questions fréquentes, ou utilisez les options de contact ci-dessous si vous avez besoin d’aide.",
+    "title": "Contact",
+    "metaDescription": "Réponses et options de contact pour Peels.",
+    "subtitle": "Écrivez-nous ci-dessus si vous avez besoin d’aide, ou parcourez les FAQ ci-dessous pour les questions courantes.",
     "peelsFaq": {
       "title": "À propos de Peels",
       "communityTitle": "Communauté et partenaires",
@@ -639,16 +639,8 @@
     }
   },
   "Contact": {
-    "title": "Nous contacter",
-    "subtitle": "Voici comment joindre l’équipe Peels.",
-    "contactLabel": "Si vous voulez",
-    "contactOptions": {
-      "general": "Poser une question générale",
-      "support": "Obtenir de l’aide",
-      "dw": "Parler à Danny",
-      "newsletter": "Parler de l’infolettre"
-    },
-    "emailLabel": "Le mieux est d’écrire à",
+    "title": "Contact",
+    "emailLabel": "Écrivez-nous à",
     "copyButton": {
       "copied": "Copié !",
       "copying": "Copie en cours...",
