+++ src/app/(core)/(static)/contact/page.tsx
@@ -1,20 +1,71 @@
-import { permanentRedirect } from "next/navigation";
-
+import { useTranslations } from "next-intl";
+import { getTranslations } from "next-intl/server";
 import { siteConfig } from "@/config/site";
-import {
-  serialiseSearchParams,
-  type StaticPageSearchParams,
-} from "@/utils/searchParams";
+import StaticPageMain from "@/components/StaticPageMain";
+import StaticPageHeader from "@/components/StaticPageHeader";
+import StaticPageSection from "@/components/StaticPageSection";
+import SupportFaq from "@/components/SupportFaq";
+import PeelsFaq from "@/components/PeelsFaq";
+import HeaderBlock from "@/components/HeaderBlock";
+import ContactEmail from "@/components/ContactEmail/ContactEmail";
+import { HelpFaqJsonLd } from "@/components/FaqJsonLd/FaqJsonLd";
+import { createPeelsMetadata } from "@/utils/seo";
 
-type ContactPageProps = {
-  searchParams?: Promise<StaticPageSearchParams>;
-};
+export async function generateMetadata() {
+  const t = await getTranslations("Support");
+  const description = t("metaDescription");
 
-export default async function Contact({ searchParams }: ContactPageProps) {
-  const params = await searchParams;
-  const queryString = serialiseSearchParams(params);
+  const title = t("title");
 
-  permanentRedirect(
-    `${siteConfig.links.help}${queryString ? `?${queryString}` : ""}#contact`
+  return createPeelsMetadata({
+    canonicalPath: "/contact",
+    title,
+    description,
+    openGraph: {
+      title: `${title} · ${siteConfig.name}`,
+      description,
+    },
+    twitter: {
+      title: `${title} · ${siteConfig.name}`,
+      description,
+    },
+  });
+}
+
+export default function Contact() {
+  const t = useTranslations("Support");
+  return (
+    <StaticPageMain>
+      <HelpFaqJsonLd />
+      <StaticPageHeader title={t("title")} subtitle={t("subtitle")} />
+
+      <StaticPageSection id="contact" padding={null}>
+        <HeaderBlock>
+          <p>{t("contact.subtitle")}</p>
+        </HeaderBlock>
+        <ContactEmail />
+      </StaticPageSection>
+
+      <StaticPageSection>
+        <HeaderBlock>
+          <h2>{t("supportFaq.title")}</h2>
+        </HeaderBlock>
+        <SupportFaq />
+      </StaticPageSection>
+
+      <StaticPageSection>
+        <HeaderBlock>
+          <h2>{t("peelsFaq.title")}</h2>
+        </HeaderBlock>
+        <PeelsFaq variant="about" />
+      </StaticPageSection>
+
+      <StaticPageSection>
+        <HeaderBlock>
+          <h2>{t("peelsFaq.communityTitle")}</h2>
+        </HeaderBlock>
+        <PeelsFaq variant="community" />
+      </StaticPageSection>
+    </StaticPageMain>
   );
 }
