+++ src/app/(core)/(static)/help/page.tsx
@@ -1,75 +1,20 @@
-import { Suspense } from "react";
-import { useTranslations } from "next-intl";
-import { getTranslations } from "next-intl/server";
+import { permanentRedirect } from "next/navigation";
+
 import { siteConfig } from "@/config/site";
-import StaticPageMain from "@/components/StaticPageMain";
-import StaticPageHeader from "@/components/StaticPageHeader";
-import StaticPageSection from "@/components/StaticPageSection";
-import SupportFaq from "@/components/SupportFaq";
-import PeelsFaq from "@/components/PeelsFaq";
-import HeaderBlock from "@/components/HeaderBlock";
-import EmailSelector from "@/components/EmailSelector/EmailSelector";
-import { HelpFaqJsonLd } from "@/components/FaqJsonLd/FaqJsonLd";
-import { createPeelsMetadata } from "@/utils/seo";
+import {
+  serialiseSearchParams,
+  type StaticPageSearchParams,
+} from "@/utils/searchParams";
 
-export async function generateMetadata() {
-  const t = await getTranslations("Support");
-  const description = t("metaDescription");
+type HelpPageProps = {
+  searchParams?: Promise<StaticPageSearchParams>;
+};
 
-  const title = t("title");
+export default async function Help({ searchParams }: HelpPageProps) {
+  const params = await searchParams;
+  const queryString = serialiseSearchParams(params);
 
-  return createPeelsMetadata({
-    canonicalPath: "/help",
-    title,
-    description,
-    openGraph: {
-      title: `${title} · ${siteConfig.name}`,
-      description,
-    },
-    twitter: {
-      title: `${title} · ${siteConfig.name}`,
-      description,
-    },
-  });
-}
-
-export default function Help() {
-  const t = useTranslations("Support");
-  return (
-    <StaticPageMain>
-      <HelpFaqJsonLd />
-      <StaticPageHeader title={t("title")} subtitle={t("subtitle")} />
-
-      <StaticPageSection padding={null}>
-        <HeaderBlock>
-          <h2>{t("supportFaq.title")}</h2>
-        </HeaderBlock>
-        <SupportFaq />
-      </StaticPageSection>
-
-      <StaticPageSection>
-        <HeaderBlock>
-          <h2>{t("peelsFaq.title")}</h2>
-        </HeaderBlock>
-        <PeelsFaq variant="about" />
-      </StaticPageSection>
-
-      <StaticPageSection>
-        <HeaderBlock>
-          <h2>{t("peelsFaq.communityTitle")}</h2>
-        </HeaderBlock>
-        <PeelsFaq variant="community" />
-      </StaticPageSection>
-
-      <StaticPageSection id="contact">
-        <HeaderBlock>
-          <h2>{t("contact.title")}</h2>
-          <p>{t("contact.subtitle")}</p>
-        </HeaderBlock>
-        <Suspense>
-          <EmailSelector />
-        </Suspense>
-      </StaticPageSection>
-    </StaticPageMain>
+  permanentRedirect(
+    `${siteConfig.links.contact}${queryString ? `?${queryString}` : ""}`
   );
 }
