+++ src/app/sitemap.ts
@@ -38,7 +38,7 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
       priority: 0.9,
     },
     {
-      url: `${siteConfig.url}/help`,
+      url: `${siteConfig.url}/contact`,
       lastModified: new Date(),
       changeFrequency: "monthly",
       priority: 0.7,
