+++ src/components/DecodedSpan/DecodedSpan.tsx
@@ -1,9 +1,10 @@
 "use client";
 
 import type { PropsWithChildren } from "react";
+import { decodeEncodedEmail } from "@/utils/email";
 
 function DecodedSpan({ children }: PropsWithChildren<{ children: string }>) {
-  return <span>{atob(children)}</span>;
+  return <span>{decodeEncodedEmail(children)}</span>;
 }
 
 export default DecodedSpan;
