--- src/components/EmailSelector/index.ts
@@ -1,2 +1,0 @@
-export * from "./EmailSelector";
-export { default } from "./EmailSelector";
