+++ src/content/legal/de/privacy.mdx
@@ -85,7 +85,7 @@ Du hast in Bezug auf deine personenbezogenen Daten die folgenden Rechte:
 5. **Datenübertragbarkeit**: deine Daten in einem portablen Format anfordern.
 6. **Widerspruch**: der Verarbeitung deiner personenbezogenen Daten in bestimmten Situationen widersprechen.
 
-Die meisten dieser Rechte kannst du direkt über deine Profileinstellungen ausüben, wo du eine vollständige Kopie deiner personenbezogenen Daten herunterladen, deine Informationen aktualisieren oder dein Konto löschen kannst. Bitte <EncodedEmailLink address={siteConfig.encodedEmail.support}>kontaktiere uns</EncodedEmailLink> für weitere Anfragen oder wenn du Unterstützung benötigst. Für bestimmte Anfragen kann eine Identitätsprüfung erforderlich sein.
+Die meisten dieser Rechte kannst du direkt über deine Profileinstellungen ausüben, wo du eine vollständige Kopie deiner personenbezogenen Daten herunterladen, deine Informationen aktualisieren oder dein Konto löschen kannst. Bitte <EncodedEmailLink address={siteConfig.encodedEmail.team}>kontaktiere uns</EncodedEmailLink> für weitere Anfragen oder wenn du Unterstützung benötigst. Für bestimmte Anfragen kann eine Identitätsprüfung erforderlich sein.
 
 ## 5. Wie lange wir deine Daten speichern
 
@@ -106,4 +106,4 @@ Wir können diese Datenschutzerklärung von Zeit zu Zeit aktualisieren, um Ände
 
 ## 8. Kontakt
 
-Bitte <EncodedEmailLink address={siteConfig.encodedEmail.support}>kontaktiere uns</EncodedEmailLink>, wenn du Fragen oder Bedenken zu dieser Datenschutzerklärung hast oder deine Rechte ausüben möchtest.
+Bitte <EncodedEmailLink address={siteConfig.encodedEmail.team}>kontaktiere uns</EncodedEmailLink>, wenn du Fragen oder Bedenken zu dieser Datenschutzerklärung hast oder deine Rechte ausüben möchtest.
