+++ src/content/legal/de/terms.mdx
@@ -84,7 +84,7 @@ Du hast in Bezug auf deine personenbezogenen Daten die folgenden Rechte:
 5. **Recht auf Datenübertragbarkeit**: Du kannst deine Daten in einem portablen Format anfordern.
 6. **Recht auf Widerspruch**: Du kannst unserer Nutzung deiner personenbezogenen Daten in bestimmten Situationen widersprechen, auch für Direktmarketing.
 
-Die meisten dieser Rechte kannst du direkt über deine Profileinstellungen ausüben, wo du eine vollständige Kopie deiner personenbezogenen Daten herunterladen, deine Informationen aktualisieren oder dein Konto löschen kannst. Bitte <EncodedEmailLink address={siteConfig.encodedEmail.support}>kontaktiere uns</EncodedEmailLink> für weitere Anfragen oder Unterstützung. Für bestimmte Anfragen kann eine Identitätsprüfung erforderlich sein.
+Die meisten dieser Rechte kannst du direkt über deine Profileinstellungen ausüben, wo du eine vollständige Kopie deiner personenbezogenen Daten herunterladen, deine Informationen aktualisieren oder dein Konto löschen kannst. Bitte <EncodedEmailLink address={siteConfig.encodedEmail.team}>kontaktiere uns</EncodedEmailLink> für weitere Anfragen oder Unterstützung. Für bestimmte Anfragen kann eine Identitätsprüfung erforderlich sein.
 
 Daten können aufgrund der von uns genutzten Dienste in Länder außerhalb deines Standorts übertragen oder dort verarbeitet werden (z. B. in den USA). Wir stellen sicher, dass solche Übermittlungen den geltenden Datenschutzgesetzen entsprechen.
 
@@ -100,7 +100,7 @@ Wir können diese Bedingungen von Zeit zu Zeit aktualisieren, um Änderungen an 
 
 ## 9. Beschwerden und Meldung von Problemen
 
-Wenn du auf ein Problem stößt oder einen Verstoß gegen diese Bedingungen melden möchtest, <EncodedEmailLink address={siteConfig.encodedEmail.support}>kontaktiere uns</EncodedEmailLink>. Wir werden Probleme innerhalb eines angemessenen Zeitraums untersuchen und bearbeiten, sind aber nicht verpflichtet, Streitigkeiten zwischen Nutzerinnen und Nutzern zu schlichten.
+Wenn du auf ein Problem stößt oder einen Verstoß gegen diese Bedingungen melden möchtest, <EncodedEmailLink address={siteConfig.encodedEmail.team}>kontaktiere uns</EncodedEmailLink>. Wir werden Probleme innerhalb eines angemessenen Zeitraums untersuchen und bearbeiten, sind aber nicht verpflichtet, Streitigkeiten zwischen Nutzerinnen und Nutzern zu schlichten.
 
 ## 10. Sperrung und Kündigung von Konten
 
@@ -116,4 +116,4 @@ Diese Bedingungen unterliegen dem Recht Australiens. Durch die Nutzung von Peels
 
 ## 12. Kontakt
 
-Bitte <EncodedEmailLink address={siteConfig.encodedEmail.support}>kontaktiere uns</EncodedEmailLink>, wenn du Fragen zu diesen Bedingungen hast.
+Bitte <EncodedEmailLink address={siteConfig.encodedEmail.team}>kontaktiere uns</EncodedEmailLink>, wenn du Fragen zu diesen Bedingungen hast.
