+++ src/content/legal/es/privacy.mdx
@@ -85,7 +85,7 @@ Tienes los siguientes derechos con respecto a tus datos personales:
 5. **Portabilidad**: solicitar tus datos en un formato portable.
 6. **Oposición**: oponerte a la forma en que tratamos tus datos personales en determinadas situaciones.
 
-La mayoría de estos derechos pueden ejercerse directamente desde la configuración de tu perfil, donde puedes descargar una copia completa de tus datos personales, actualizar tu información o eliminar tu cuenta. <EncodedEmailLink address={siteConfig.encodedEmail.support}>Contáctanos</EncodedEmailLink> para cualquier otra solicitud o si necesitas ayuda. En determinados casos puede requerirse la verificación de tu identidad.
+La mayoría de estos derechos pueden ejercerse directamente desde la configuración de tu perfil, donde puedes descargar una copia completa de tus datos personales, actualizar tu información o eliminar tu cuenta. <EncodedEmailLink address={siteConfig.encodedEmail.team}>Contáctanos</EncodedEmailLink> para cualquier otra solicitud o si necesitas ayuda. En determinados casos puede requerirse la verificación de tu identidad.
 
 ## 5. Cuánto tiempo conservamos tus datos
 
@@ -107,4 +107,4 @@ Podemos actualizar esta política de privacidad periódicamente para reflejar ca
 ## 8. Contacto
 
 {/* prettier-ignore */}
-<EncodedEmailLink address={siteConfig.encodedEmail.support}>Contáctanos</EncodedEmailLink> si tienes preguntas o inquietudes sobre esta política de privacidad, o si deseas ejercer tus derechos.
+<EncodedEmailLink address={siteConfig.encodedEmail.team}>Contáctanos</EncodedEmailLink> si tienes preguntas o inquietudes sobre esta política de privacidad, o si deseas ejercer tus derechos.
