+++ src/content/legal/es/terms.mdx
@@ -83,7 +83,7 @@ Tienes los siguientes derechos con respecto a tus datos personales:
 5. **Derecho a la portabilidad de los datos**: puedes solicitar tus datos en un formato portable.
 6. **Derecho de oposición**: puedes oponerte al uso de tus datos personales en determinadas situaciones, incluido el marketing directo.
 
-La mayoría de estos derechos pueden ejercerse directamente desde la configuración de tu perfil, donde puedes descargar una copia completa de tus datos personales, actualizar tu información o eliminar tu cuenta. <EncodedEmailLink address={siteConfig.encodedEmail.support}>Contáctanos</EncodedEmailLink> para cualquier otra solicitud o ayuda. En determinados casos puede requerirse la verificación de tu identidad.
+La mayoría de estos derechos pueden ejercerse directamente desde la configuración de tu perfil, donde puedes descargar una copia completa de tus datos personales, actualizar tu información o eliminar tu cuenta. <EncodedEmailLink address={siteConfig.encodedEmail.team}>Contáctanos</EncodedEmailLink> para cualquier otra solicitud o ayuda. En determinados casos puede requerirse la verificación de tu identidad.
 
 Los datos pueden transferirse o procesarse en países fuera de tu ubicación (por ejemplo, Estados Unidos) debido a los servicios que utilizamos. Nos aseguramos de que dichas transferencias cumplan con la legislación aplicable en materia de protección de datos.
 
@@ -99,7 +99,7 @@ Podemos actualizar estos Términos periódicamente para reflejar cambios en nues
 
 ## 9. Quejas y notificación de problemas
 
-Si encuentras un problema o necesitas informar de una infracción de estos Términos, <EncodedEmailLink address={siteConfig.encodedEmail.support}>contáctanos</EncodedEmailLink>. Investigaremos y abordaremos los problemas en un plazo razonable, pero no estamos obligados a mediar disputas entre personas usuarias.
+Si encuentras un problema o necesitas informar de una infracción de estos Términos, <EncodedEmailLink address={siteConfig.encodedEmail.team}>contáctanos</EncodedEmailLink>. Investigaremos y abordaremos los problemas en un plazo razonable, pero no estamos obligados a mediar disputas entre personas usuarias.
 
 ## 10. Suspensión y cancelación de cuenta
 
@@ -116,4 +116,4 @@ Estos Términos se rigen por las leyes de Australia. Al usar Peels, aceptas some
 ## 12. Contacto
 
 {/* prettier-ignore */}
-<EncodedEmailLink address={siteConfig.encodedEmail.support}>Contáctanos</EncodedEmailLink> si tienes preguntas sobre estos Términos.
+<EncodedEmailLink address={siteConfig.encodedEmail.team}>Contáctanos</EncodedEmailLink> si tienes preguntas sobre estos Términos.
