+++ src/content/legal/fr/privacy.mdx
@@ -85,7 +85,7 @@ Vous disposez des droits suivants concernant vos données personnelles :
 5. **Portabilité** : demander vos données dans un format portable.
 6. **Opposition** : vous opposer à la manière dont nous traitons vos données personnelles dans certaines situations.
 
-La plupart de ces droits peuvent être exercés directement depuis vos paramètres de profil, où vous pouvez télécharger une copie complète de vos données personnelles, mettre à jour vos informations ou supprimer votre compte. Veuillez <EncodedEmailLink address={siteConfig.encodedEmail.support}>nous contacter</EncodedEmailLink> pour toute autre demande ou si vous avez besoin d’aide. Une vérification d’identité peut être requise pour certaines demandes.
+La plupart de ces droits peuvent être exercés directement depuis vos paramètres de profil, où vous pouvez télécharger une copie complète de vos données personnelles, mettre à jour vos informations ou supprimer votre compte. Veuillez <EncodedEmailLink address={siteConfig.encodedEmail.team}>nous contacter</EncodedEmailLink> pour toute autre demande ou si vous avez besoin d’aide. Une vérification d’identité peut être requise pour certaines demandes.
 
 ## 5. Combien de temps nous conservons vos données
 
@@ -106,4 +106,4 @@ Nous pouvons mettre à jour cette politique de confidentialité de temps à autr
 
 ## 8. Contact
 
-Veuillez <EncodedEmailLink address={siteConfig.encodedEmail.support}>nous contacter</EncodedEmailLink> si vous avez des questions ou des préoccupations concernant cette politique de confidentialité, ou si vous souhaitez exercer vos droits.
+Veuillez <EncodedEmailLink address={siteConfig.encodedEmail.team}>nous contacter</EncodedEmailLink> si vous avez des questions ou des préoccupations concernant cette politique de confidentialité, ou si vous souhaitez exercer vos droits.
