+++ src/content/legal/fr/terms.mdx
@@ -84,7 +84,7 @@ Vous disposez des droits suivants concernant vos données personnelles :
 5. **Droit à la portabilité des données** : vous pouvez demander que vos données vous soient fournies dans un format portable.
 6. **Droit d’opposition** : vous pouvez vous opposer à l’utilisation de vos données personnelles dans certaines situations, y compris à des fins de marketing direct.
 
-La plupart de ces droits peuvent être exercés directement dans vos paramètres de profil, où vous pouvez télécharger une copie complète de vos données personnelles, mettre à jour vos informations ou supprimer votre compte. Veuillez <EncodedEmailLink address={siteConfig.encodedEmail.support}>nous contacter</EncodedEmailLink> pour toute autre demande ou assistance. Une vérification d’identité peut être exigée pour certaines demandes.
+La plupart de ces droits peuvent être exercés directement dans vos paramètres de profil, où vous pouvez télécharger une copie complète de vos données personnelles, mettre à jour vos informations ou supprimer votre compte. Veuillez <EncodedEmailLink address={siteConfig.encodedEmail.team}>nous contacter</EncodedEmailLink> pour toute autre demande ou assistance. Une vérification d’identité peut être exigée pour certaines demandes.
 
 Les données peuvent être transférées ou traitées dans des pays situés hors de votre lieu de résidence (par exemple aux États-Unis) en raison des services que nous utilisons. Nous veillons à ce que ces transferts respectent les lois applicables en matière de protection des données.
 
@@ -100,7 +100,7 @@ Nous pouvons mettre à jour ces Conditions de temps à autre afin de refléter d
 
 ## 9. Plaintes et signalement de problèmes
 
-Si vous rencontrez un problème ou souhaitez signaler une violation des présentes Conditions, veuillez <EncodedEmailLink address={siteConfig.encodedEmail.support}>nous contacter</EncodedEmailLink>. Nous enquêterons et traiterons les problèmes dans un délai raisonnable, sans être tenus de servir de médiateur dans les litiges entre personnes utilisatrices.
+Si vous rencontrez un problème ou souhaitez signaler une violation des présentes Conditions, veuillez <EncodedEmailLink address={siteConfig.encodedEmail.team}>nous contacter</EncodedEmailLink>. Nous enquêterons et traiterons les problèmes dans un délai raisonnable, sans être tenus de servir de médiateur dans les litiges entre personnes utilisatrices.
 
 ## 10. Suspension et résiliation du compte
 
@@ -116,4 +116,4 @@ Les présentes Conditions sont régies par le droit australien. En utilisant Pee
 
 ## 12. Contact
 
-Veuillez <EncodedEmailLink address={siteConfig.encodedEmail.support}>nous contacter</EncodedEmailLink> si vous avez des questions sur ces Conditions.
+Veuillez <EncodedEmailLink address={siteConfig.encodedEmail.team}>nous contacter</EncodedEmailLink> si vous avez des questions sur ces Conditions.
