+++ src/content/legal/privacy.mdx
@@ -85,7 +85,7 @@ You have the following rights regarding your personal data:
 5. **Portability**: Request your data in a portable format.
 6. **Objection**: Object to how we process your personal data in specific situations.
 
-Most of these rights can be exercised directly through your profile settings, where you can download a complete copy of your personal data, update your information, or delete your account. Please <EncodedEmailLink address={siteConfig.encodedEmail.support}>contact us</EncodedEmailLink> for other requests or assistance. Identity verification may be required for certain requests.
+Most of these rights can be exercised directly through your profile settings, where you can download a complete copy of your personal data, update your information, or delete your account. Please <EncodedEmailLink address={siteConfig.encodedEmail.team}>contact us</EncodedEmailLink> for other requests or assistance. Identity verification may be required for certain requests.
 
 ## 5. How long we keep your data
 
@@ -106,4 +106,4 @@ We may update this privacy policy from time to time to reflect changes to our se
 
 ## 8. Contact
 
-Please <EncodedEmailLink address={siteConfig.encodedEmail.support}>contact us</EncodedEmailLink> if you have questions or concerns about this privacy policy, or wish to exercise your rights.
+Please <EncodedEmailLink address={siteConfig.encodedEmail.team}>contact us</EncodedEmailLink> if you have questions or concerns about this privacy policy, or wish to exercise your rights.
