+++ src/content/legal/pt-BR/privacy.mdx
@@ -85,7 +85,7 @@ Você tem os seguintes direitos em relação aos seus dados pessoais:
 5. **Portabilidade**: solicitar os seus dados em um formato portátil.
 6. **Oposição**: opor-se à forma como processamos os seus dados pessoais em determinadas situações.
 
-A maior parte desses direitos pode ser exercida diretamente nas configurações do seu perfil, onde você pode baixar uma cópia completa dos seus dados pessoais, atualizar as suas informações ou excluir a sua conta. <EncodedEmailLink address={siteConfig.encodedEmail.support}>Entre em contato conosco</EncodedEmailLink> para outros pedidos ou caso precise de ajuda. A verificação de identidade pode ser exigida em determinados casos.
+A maior parte desses direitos pode ser exercida diretamente nas configurações do seu perfil, onde você pode baixar uma cópia completa dos seus dados pessoais, atualizar as suas informações ou excluir a sua conta. <EncodedEmailLink address={siteConfig.encodedEmail.team}>Entre em contato conosco</EncodedEmailLink> para outros pedidos ou caso precise de ajuda. A verificação de identidade pode ser exigida em determinados casos.
 
 ## 5. Por quanto tempo mantemos os seus dados
 
@@ -107,4 +107,4 @@ Podemos atualizar esta política de privacidade de tempos em tempos para refleti
 ## 8. Contato
 
 {/* prettier-ignore */}
-<EncodedEmailLink address={siteConfig.encodedEmail.support}>Entre em contato conosco</EncodedEmailLink> se você tiver dúvidas ou preocupações sobre esta política de privacidade, ou se quiser exercer os seus direitos.
+<EncodedEmailLink address={siteConfig.encodedEmail.team}>Entre em contato conosco</EncodedEmailLink> se você tiver dúvidas ou preocupações sobre esta política de privacidade, ou se quiser exercer os seus direitos.
