+++ src/content/legal/pt-BR/terms.mdx
@@ -84,7 +84,7 @@ Você tem os seguintes direitos em relação aos seus dados pessoais:
 5. **Direito à portabilidade dos dados**: você pode solicitar os seus dados em formato portátil.
 6. **Direito de oposição**: você pode se opor ao uso dos seus dados pessoais em determinadas situações, inclusive para fins de marketing direto.
 
-A maior parte desses direitos pode ser exercida diretamente nas configurações do seu perfil, onde você pode baixar uma cópia completa dos seus dados pessoais, atualizar as suas informações ou excluir a sua conta. <EncodedEmailLink address={siteConfig.encodedEmail.support}>Entre em contato conosco</EncodedEmailLink> para outros pedidos ou assistência. A verificação de identidade pode ser exigida em alguns casos.
+A maior parte desses direitos pode ser exercida diretamente nas configurações do seu perfil, onde você pode baixar uma cópia completa dos seus dados pessoais, atualizar as suas informações ou excluir a sua conta. <EncodedEmailLink address={siteConfig.encodedEmail.team}>Entre em contato conosco</EncodedEmailLink> para outros pedidos ou assistência. A verificação de identidade pode ser exigida em alguns casos.
 
 Os dados podem ser transferidos ou processados em países fora da sua localização (por exemplo os Estados Unidos) devido aos serviços que utilizamos. Garantimos que essas transferências cumpram as leis aplicáveis de proteção de dados.
 
@@ -100,7 +100,7 @@ Podemos atualizar estes Termos de tempos em tempos para refletir mudanças nos n
 
 ## 9. Reclamações e reporte de problemas
 
-Se você tiver um problema ou quiser relatar uma violação destes Termos, <EncodedEmailLink address={siteConfig.encodedEmail.support}>entre em contato conosco</EncodedEmailLink>. Investigaremos e trataremos os problemas dentro de um prazo razoável, mas não somos obrigados a mediar disputas entre pessoas usuárias.
+Se você tiver um problema ou quiser relatar uma violação destes Termos, <EncodedEmailLink address={siteConfig.encodedEmail.team}>entre em contato conosco</EncodedEmailLink>. Investigaremos e trataremos os problemas dentro de um prazo razoável, mas não somos obrigados a mediar disputas entre pessoas usuárias.
 
 ## 10. Suspensão e encerramento de conta
 
@@ -117,4 +117,4 @@ Estes Termos são regidos pelas leis da Austrália. Ao usar o Peels, você conco
 ## 12. Contato
 
 {/* prettier-ignore */}
-<EncodedEmailLink address={siteConfig.encodedEmail.support}>Entre em contato conosco</EncodedEmailLink> se você tiver qualquer dúvida sobre estes Termos.
+<EncodedEmailLink address={siteConfig.encodedEmail.team}>Entre em contato conosco</EncodedEmailLink> se você tiver qualquer dúvida sobre estes Termos.
