+++ src/content/legal/terms.mdx
@@ -83,7 +83,7 @@ You have the following rights regarding your personal data:
 5. **Right to data portability**: You can request that we provide your data in a portable format.
 6. **Right to object**: You can object to our use of your personal data in certain situations, including for direct marketing purposes.
 
-Most of these rights can be exercised directly through your profile settings, where you can download a complete copy of your personal data, update your information, or delete your account. Please <EncodedEmailLink address={siteConfig.encodedEmail.support}>contact us</EncodedEmailLink> for other requests or assistance. Identity verification may be required for certain requests.
+Most of these rights can be exercised directly through your profile settings, where you can download a complete copy of your personal data, update your information, or delete your account. Please <EncodedEmailLink address={siteConfig.encodedEmail.team}>contact us</EncodedEmailLink> for other requests or assistance. Identity verification may be required for certain requests.
 
 Data may be transferred to or processed in countries outside your location (e.g., the United States) due to the services we use (e.g., Vercel). We ensure such transfers comply with applicable data protection laws.
 
@@ -99,7 +99,7 @@ We may update these Terms from time to time to reflect changes in our services, 
 
 ## 9. Complaints and reporting issues
 
-If you encounter a problem or need to report a violation of these Terms, please <EncodedEmailLink address={siteConfig.encodedEmail.support}>contact us</EncodedEmailLink>. We will investigate and address issues in a timely manner but are not obligated to mediate disputes between users.
+If you encounter a problem or need to report a violation of these Terms, please <EncodedEmailLink address={siteConfig.encodedEmail.team}>contact us</EncodedEmailLink>. We will investigate and address issues in a timely manner but are not obligated to mediate disputes between users.
 
 ## 10. Account suspension and termination
 
@@ -115,4 +115,4 @@ These Terms are governed by the laws of Australia. By using Peels, you agree to 
 
 ## 12. Contact
 
-Please <EncodedEmailLink address={siteConfig.encodedEmail.support}>contact us</EncodedEmailLink> if you have any questions about these Terms.
+Please <EncodedEmailLink address={siteConfig.encodedEmail.team}>contact us</EncodedEmailLink> if you have any questions about these Terms.
