+++ src/utils/email.ts
@@ -1,0 +1,3 @@
+export function decodeEncodedEmail(encoded: string): string {
+  return atob(encoded);
+}
